src/requirements.txt
customtkinter>=5.2.0
pyAesCrypt>=6.1.0
pyperclip>=1.8.0
Pillow>=10.0.0

